removed MeshRelation
