+++ bindings/c/include/conv.h
@@ -14,7 +14,6 @@ ManifoldBox *to_c(manifold::Box *m);
 ManifoldCurvature *to_c(manifold::Curvature *m);
 ManifoldComponents *to_c(manifold::Components *components);
 ManifoldError to_c(manifold::Manifold::Error error);
-ManifoldMeshRelation *to_c(manifold::MeshRelation *m);
 ManifoldMaterial *to_c(manifold::Material *m);
 ManifoldExportOptions *to_c(manifold::ExportOptions *m);
 ManifoldVec3 to_c(glm::vec3 v);
@@ -29,7 +28,6 @@ const manifold::MeshGL *from_c(ManifoldMeshGL *m);
 const manifold::Box *from_c(ManifoldBox *m);
 const manifold::Curvature *from_c(ManifoldCurvature *m);
 const manifold::Components *from_c(ManifoldComponents *components);
-const manifold::MeshRelation *from_c(ManifoldMeshRelation *m);
 manifold::Material *from_c(ManifoldMaterial *mat);
 manifold::ExportOptions *from_c(ManifoldExportOptions *options);
 glm::vec3 from_c(ManifoldVec3 v);
