+++ bindings/wasm/manifold.d.ts
@@ -30,11 +30,6 @@ type Properties = {
   surfaceArea: number,
   volume: number
 };
-type TriRef = {
-  meshID: number,
-  originalID: number,
-  tri: number
-};
 type Curvature = {
   maxMeanCurvature: number,
   minMeanCurvature: number,
@@ -42,9 +37,6 @@ type Curvature = {
   minGaussianCurvature: number,
   vertMeanCurvature: number[],
   vertGaussianCurvature: number[]
-};
-type MeshRelation = {
-  triRef: TriRef[]
 };
 
 declare class Mesh {
@@ -222,23 +214,6 @@ declare class Manifold {
   getMesh(): Mesh;
 
   /**
-   * Gets the relationship to the previous meshes, for the purpose of assigning
-   * properties like texture coordinates. The triRef vector is the same length
-   * as Mesh.triVerts: TriRef.originalID indicates the source mesh and
-   * TriRef.tri is that mesh's triangle index to which these barycentric
-   * coordinates refer. TriRef.vertBary gives an index for each vertex into the
-   * barycentric vector if that index is >= 0, indicating it is a new vertex. If
-   * the index is < 0, this indicates it is an original vertex, the index + 3
-   * vert of the referenced triangle.
-   *
-   * TriRef.meshID is a unique ID to the particular instance of a given mesh.
-   * For instance, if you want to convert the triangle mesh to a polygon mesh,
-   * all the triangles from a given face will have the same .meshID and .tri
-   * values.
-   */
-  getMeshRelation(): MeshRelation;
-
-  /**
    * If you copy a manifold, but you want this new copy to have new properties
    * (e.g. a different UV mapping), you can reset its meshIDs to a new original,
    * meaning it will now be referenced by its descendants instead of the meshes
@@ -255,7 +230,7 @@ declare class Manifold {
 
   /**
    * If this mesh is an original, this returns its meshID that can be referenced
-   * by product manifolds' MeshRelation. If this manifold is a product, this
+   * by product manifolds. If this manifold is a product, this
    * returns -1.
    */
   originalID(): number;
