+++ src/manifold/src/manifold.cpp
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <algorithm>
 #include <map>
 
 #include "boolean3.h"
@@ -404,24 +405,6 @@ Properties Manifold::GetProperties() const {
  */
 Curvature Manifold::GetCurvature() const {
   return GetCsgLeafNode().GetImpl()->GetCurvature();
-}
-
-/**
- * Gets the relationship to the previous meshes, for the purpose of assigning
- * properties like texture coordinates. The triRef vector is the same length as
- * Mesh.triVerts: TriRef.originalID indicates the source mesh and TriRef.tri
- * is that mesh's triangle index of which this triangle is a part.
- *
- * TriRef.meshID is a unique ID to the particular instance of a given mesh. For
- * instance, if you want to convert the triangle mesh to a polygon mesh, all the
- * triangles from a given face will have the same .meshID and .tri values.
- */
-MeshRelation Manifold::GetMeshRelation() const {
-  MeshRelation out;
-  const auto& relation = GetCsgLeafNode().GetImpl()->meshRelation_;
-  out.triRef.insert(out.triRef.end(), relation.triRef.begin(),
-                    relation.triRef.end());
-  return out;
 }
 
 /**
