+++ src/utilities/include/public.h
@@ -194,7 +194,8 @@ struct MeshGL {
   /// Optional: Indicates runs of triangles that correspond to a particular
   /// meshID. The runs encompass all of triVerts and are sorted
   /// lexicographically by originalID, then meshID. Run i begins at
-  /// triVerts[runIndex[i]] and ends at triVerts[runIndex[i+1]].
+  /// triVerts[runIndex[i]] and ends at triVerts[runIndex[i+1]]. All runIndex
+  /// values are divisible by 3.
   std::vector<uint32_t> runIndex;
   /// Optional: The OriginalID of the mesh this triangle run came from. This ID
   /// is ideal for reapplying materials to the output mesh.
@@ -269,35 +270,7 @@ struct Components {
   std::vector<int> indices;
   int numComponents;
 };
-
-/**
- * Part of MeshRelation - represents a single triangle relation to an original
- * Mesh.
- */
-struct TriRef {
-  /// The unique ID of the mesh instance of this triangle. If .meshID and .tri
-  /// match for two triangles, then they are coplanar and came from the same
-  /// face.
-  int meshID;
-  /// The OriginalID of the mesh this triangle came from. This ID is ideal for
-  /// reapplying properties like UV coordinates to the output mesh.
-  int originalID;
-  /// The triangle index of the original triangle this was part of:
-  /// Mesh.triVerts[tri].
-  int tri;
-};
 /** @} */
-
-/**
- *  @ingroup Connections
- *  Represents the relationship of this output Mesh to all input Meshes that
- * eventually led to it, see Manifold.GetMeshRelation().
- */
-struct MeshRelation {
-  /// A vector matching Mesh.triVerts that contains the relation of each output
-  /// triangle to a single input triangle.
-  std::vector<TriRef> triRef;
-};
 
 /**
  * @ingroup Connections
@@ -529,11 +502,6 @@ inline std::ostream& operator<<(std::ostream& stream, const glm::mat4x3& mat) {
   return stream << tam[0] << std::endl
                 << tam[1] << std::endl
                 << tam[2] << std::endl;
-}
-
-inline std::ostream& operator<<(std::ostream& stream, const TriRef& ref) {
-  return stream << "meshID: " << ref.meshID
-                << ", originalID: " << ref.originalID << ", tri: " << ref.tri;
 }
 
 /**
